feat: add Dashboard page
Return a DashboardContainer that has 3 Custom Cards as childrens. To hydratate data, is called 'getServerSideProps' to get info about user profile. If no token was not found in cookies storage, the user is redirect to the homepage. After that, the useEffect update states of profile and contacts to render dinamically user data on screen.