feat: add updateUserDate function to update an user
This function handle a little bit more the data provided. On line 49 to 64, filledData calls the cunstructor 'entries' method, then pass the 'payload' as prop to run a reduce that will check if the value is filled or not. Also, do a additional verification if the key is 'phone' to replace the masked data to a normalized string of numbers. Then, the api request is triggered calling '/profile' in URL and passing 'filledData' as props.