feat: add Accordion component and their props
The props passed are 'children', 'triggerText' and 'passwordField'. The 'children' prop is a react component that will be rendered in AccordionContent component. The 'triggerText' will be displayed in AccordionTrigger component. The 'passwordField' prop is passed in AccordionContent to a validation to render conditionaly the Custom Accordion Content or 'CustomContent' as was named.